created routes for campgrounds and reviews
